requirements.txt

#torch>=2.1.1,<3.0
#transformers #>=4.39.0,<5.0.0
asyncio>=3.4.3,<4.0
#langchain-community==0.0.29
#langchain-experimental==0.0.55
backoff
toml
#pypdf==4.3.1
ratelimit==2.2.1
loguru
pydantic==2.8.2
tenacity
rich
#Pillow==10.3.0
psutil
#sentry-sdk
python-dotenv
PyYAML
docstring_parser==0.16
black>=23.1,<25.0
ruff>=0.0.249,<0.4.5
types-toml
types-pytz>=2023.3,<2025.0
types-chardet>=5.0.4.6
mypy-protobuf>=3.0.0
pytest>=8.1.1
#pandas>=2.2.2

#networkx
aiofiles
#swarm-models
#https://github.com/jmikedupont2/swarm-models.git@main/
-e git+https://github.com/jmikedupont2/swarm-models.git@main#egg=swarm-models
#clusterops
#reportlab
#doc-master
fastapi #>=0.110.1

#openai

